scripts/requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
rdkit-pypi>=2023.3.1
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
